requirements.txt
Flask
python-dotenv==1.0.0
supabase
flask-cors
pandas